Move catanatron_server under catanatron[web]
