+++ Dockerfile.paperspace
@@ -26,9 +26,7 @@ RUN python3.8 -m pip install -r dev-requirements.txt
 #   catanatron package in the dev-requirements.txt. So we install
 #   it separately, here:
 COPY . .
-RUN python3.8 -m pip install -e catanatron
-RUN python3.8 -m pip install -e catanatron_gym
-RUN python3.8 -m pip install -e catanatron_server
+RUN python3.8 -m pip install -e catanatron[web,gym]
 RUN python3.8 -m pip install -e catanatron_experimental
 
 EXPOSE 8888
