+++ Dockerfile.web
@@ -1,3 +1,5 @@
+# Build Command:
+# docker build -t bcollazo/catanatron-server:latest . -f Dockerfile.web
 FROM python:3.12
 
 WORKDIR /app
@@ -5,15 +7,13 @@ WORKDIR /app
 RUN pip install --upgrade pip
 
 COPY . .
-RUN pip install -e catanatron
-RUN pip install -e catanatron_server
-RUN pip install -e catanatron_gym
+RUN pip install -e catanatron[web]
 RUN pip install -e catanatron_experimental
 
 EXPOSE 5000
 
 ENV FLASK_DEBUG=1
-ENV FLASK_APP=catanatron_server/catanatron_server
+ENV FLASK_APP=catanatron.web/catanatron.web
 ENV FLASK_RUN_HOST=0.0.0.0
 
 CMD flask run
