+++ catanatron/catanatron/web/wsgi.py
@@ -1,0 +1,3 @@
+from catanatron.web import create_app
+
+app = create_app()
