+++ catanatron_experimental/catanatron_experimental/cli/accumulators.py
@@ -15,8 +15,8 @@ from catanatron.state_functions import (
     get_player_buildings,
 )
 from catanatron.models.enums import VICTORY_POINT, SETTLEMENT, CITY
-from catanatron_server.models import database_session, upsert_game_state
-from catanatron_server.utils import ensure_link
+from catanatron.web.models import database_session, upsert_game_state
+from catanatron.web.utils import ensure_link
 from catanatron_experimental.utils import formatSecs
 from catanatron_experimental.machine_learning.utils import (
     get_discounted_return,
