--- catanatron_server/catanatron_server/wsgi.py
@@ -1,3 +1,0 @@
-from catanatron_server import create_app
-
-app = create_app()
