+++ dev-requirements.txt
@@ -1,8 +1,6 @@
 black==25.1.0
 blinker==1.9.0
--e catanatron
--e catanatron_gym
--e catanatron_server
+-e catanatron[web,gym]
 -e catanatron_experimental
 certifi==2025.4.26
 charset-normalizer==3.4.2
