+++ docs/requirements.txt
@@ -31,5 +31,4 @@ sphinxcontrib-qthelp==1.0.3
 sphinxcontrib-serializinghtml==1.1.5
 urllib3==1.26.6
 zipp==3.8.1
--e catanatron
--e catanatron_gym
+-e catanatron[web,gym]
