+++ docs/source/catanatron.gym.envs.rst
@@ -1,0 +1,21 @@
+catanatron.gym.envs package
+===========================
+
+Submodules
+----------
+
+catanatron.gym.envs.catanatron\_env module
+------------------------------------------
+
+.. automodule:: catanatron.gym.envs.catanatron_env
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: catanatron.gym.envs
+   :members:
+   :undoc-members:
+   :show-inheritance:
