+++ docs/source/catanatron.gym.rst
@@ -1,0 +1,29 @@
+catanatron.gym package
+======================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   catanatron.gym.envs
+
+Submodules
+----------
+
+catanatron.gym.board\_tensor\_features module
+---------------------------------------------
+
+.. automodule:: catanatron.gym.board_tensor_features
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: catanatron.gym
+   :members:
+   :undoc-members:
+   :show-inheritance:
