+++ docs/source/catanatron.web.rst
@@ -1,0 +1,45 @@
+catanatron.web package
+======================
+
+Submodules
+----------
+
+catanatron.web.api module
+-------------------------
+
+.. automodule:: catanatron.web.api
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+catanatron.web.models module
+----------------------------
+
+.. automodule:: catanatron.web.models
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+catanatron.web.utils module
+---------------------------
+
+.. automodule:: catanatron.web.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+catanatron.web.wsgi module
+--------------------------
+
+.. automodule:: catanatron.web.wsgi
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: catanatron.web
+   :members:
+   :undoc-members:
+   :show-inheritance:
