--- docs/source/catanatron_gym.envs.rst
@@ -1,21 +1,0 @@
-catanatron\_gym.envs package
-============================
-
-Submodules
-----------
-
-catanatron\_gym.envs.catanatron\_gym module
--------------------------------------------
-
-.. automodule:: catanatron_gym.envs.catanatron_gym
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: catanatron_gym.envs
-   :members:
-   :undoc-members:
-   :show-inheritance:
