--- docs/source/catanatron_gym.rst
@@ -1,37 +1,0 @@
-catanatron\_gym package
-=======================
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   catanatron_gym.envs
-
-Submodules
-----------
-
-catanatron\_gym.board\_tensor\_features module
-----------------------------------------------
-
-.. automodule:: catanatron_gym.board_tensor_features
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-catanatron\_gym.features module
--------------------------------
-
-.. automodule:: catanatron_gym.features
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: catanatron_gym
-   :members:
-   :undoc-members:
-   :show-inheritance:
