--- docs/source/catanatron_server.rst
@@ -1,45 +1,0 @@
-catanatron\_server package
-==========================
-
-Submodules
-----------
-
-catanatron\_server.api module
------------------------------
-
-.. automodule:: catanatron_server.api
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-catanatron\_server.models module
---------------------------------
-
-.. automodule:: catanatron_server.models
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-catanatron\_server.utils module
--------------------------------
-
-.. automodule:: catanatron_server.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-catanatron\_server.wsgi module
-------------------------------
-
-.. automodule:: catanatron_server.wsgi
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: catanatron_server
-   :members:
-   :undoc-members:
-   :show-inheritance:
