--- docs/source/sample.rst
@@ -1,7 +1,0 @@
-sample module
-=============
-
-.. automodule:: sample
-   :members:
-   :undoc-members:
-   :show-inheritance:
