+++ pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 testpaths = tests
-pythonpath = catanatron catanatron_server catanatron_gym catanatron_experimental
+pythonpath = catanatron catanatron_experimental
 addopts = 
     "--benchmark-columns=min,max,mean,stddev,rounds"
     --disable-warnings
