+++ requirements.txt
@@ -1,7 +1,5 @@
 blinker==1.9.0
--e catanatron
--e catanatron_gym
--e catanatron_server
+-e catanatron[gym,web]
 -e catanatron_experimental
 click==8.1.8
 cloudpickle==3.1.1
